Calendar: Move AliyahLabeller to separate file
This commit does not change any actual code.
It's prework for upcoming unit tests.
